3D-BAUSTEINE-1.2.1
First, the user interface was redesigned.
From now on, the user interface should be more flexible and,
if possible, it should also be the conditions of the target system.
The result was that the user interface now automatically adjusts
to the computer's screen resolution of the computer! The student
can then also manual changes to the interface, but these are only
valid for the current session. the current session.

Furthermore some changes were made internally.
In order to be able to make an extension beyond the "standard bodies" in
future the internal data structures as well as the flow control have been
changed.

In later versions, it should also be possible to define bodies not
previously known to the system, e.g. in a library file. e.g. in a
library file, possibly even to define bodies interactively.
(this will already possible in version 1.3.).

The flow control and body representation has been extended,
but also streamlined, so that it is more flexible, and
also smaller and faster! ( just not a Microsoft product ¯\_(ツ)_/¯ )