3D-BAUSTEINE-1.3.0-rc1
While I was still working on this thesis, I continued to work on the
3D Building Blocks project and implemented new functionality.
It was close to my heart to implement the mechanism of CSG
(Constructive Solid Geometry) as well as the possibility to define
new bodies. This has far-reaching positive consequences for the
usability of the program, because it makes it possible to create
more and more new and more complex bodies from the existing
basic bodies, as well as to insert them several times e.g. on another
position in the scene. If it is then still possible to store the newly
created bodies and read them in again, nothing more stands in the
way of the creation of object libraries. Furthermore, in version 1.3
the selection and assignment of standard textures to the bodies
(see texture.inc) has been implemented, which makes the creation
of photorealistic images much easier.